requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.0
scikit-learn==1.3.0
flask==2.3.2
flask-cors==4.0.0
webcolors==1.13
colorthief==0.2.1
scipy==1.11.1
seaborn==0.12.2
plotly==5.15.0
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0

